feat: shows the previous search result next to the current one
