+++ package.json
@@ -25,6 +25,7 @@
     "pokenode-ts": "^1.20.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-icons": "^5.3.0",
     "react-speech-recognition": "^3.10.0",
     "regenerator-runtime": "^0.14.0",
     "superjson": "^1.13.1",
