+++ src/pages/index.tsx
@@ -6,6 +6,8 @@ import { distance } from "fastest-levenshtein";
 import { api } from "~/utils/api";
 import { Footer } from "~/components/Footer";
 import Dictaphone from "~/components/Dictaphone";
+import PokemonTile from "~/components/PokemonTile";
+import { BsArrowRight } from "react-icons/bs";
 
 const title = "Carry Helper";
 const description =
@@ -13,10 +15,17 @@ const description =
 
 export default function Home() {
   const [pokemonName, setPokemonName] = useState("");
+  const [previousPokemon, setPreviousPokemon] = useState<Pokemon>(null);
+  const [tmpPokemon, setTmpPokemon] = useState<Pokemon>(null);
   const { data: pokemon, status } = api.pokemon.fetchByName.useQuery({
     name: pokemonName,
     versionGroup: "crystal",
   });
+
+  useEffect(() => {
+    setPreviousPokemon(tmpPokemon); // Save the previous Pokemon
+    !pokemon || pokemon == undefined ? null : setTmpPokemon(pokemon); // backup the current Pokemon
+  }, [pokemon]);
 
   return (
     <>
@@ -47,41 +56,19 @@ export default function Home() {
             value={pokemonName}
             onChange={(v) => setPokemonName(v)}
           />
-          {pokemonName && status === "loading" ? (
-            <p>Loading...</p>
-          ) : !pokemon || !pokemonName ? null : (
-            <div className="flex flex-col gap-2 rounded border border-white p-3 ">
-              <p className="text-center text-2xl capitalize">{pokemon.name}</p>
-              <hr />
-              <ul className="flex flex-col gap-2">
-                {Object.entries(pokemon.stats).map(([stat, value]) => (
-                  <React.Fragment key={stat}>
-                    <li className="grid grid-cols-[1fr,3ch,1fr] gap-2 capitalize">
-                      <div className="text-right">{stat}:</div>
-                      <div className="w-[3ch] text-right">{value}</div>
-                      <div
-                        className={`h-6 ${
-                          {
-                            S: "bg-green-500",
-                            A: "bg-green-300",
-                            B: "bg-yellow-400",
-                            C: "bg-yellow-600",
-                            D: "bg-red-300",
-                            F: "bg-red-500",
-                          }[statValueToTier(value)]
-                        }`}
-                        style={{ width: `${(100 / 255) * value}%` }}
-                      />
-                    </li>
-                    {stat === "special-attack" && <hr />}
-                  </React.Fragment>
-                ))}
-              </ul>
-              <hr />
-              <p className="text-lg">Moves learned on level:</p>
-              <p>{pokemon.moveLevels.join(", ")}</p>
-            </div>
-          )}
+          <section className="container flex flex-row items-center justify-center gap-12">
+            {!previousPokemon ? null : (
+              <>
+                <PokemonTile pokemon={previousPokemon} />
+                <BsArrowRight className="text-4xl" />
+              </>
+            )}
+            {pokemonName && status === "loading" ? (
+              <p>Loading...</p>
+            ) : !pokemon || !pokemonName ? null : (
+              <PokemonTile pokemon={pokemon} />
+            )}
+          </section>
         </div>
       </main>
       <Footer />
@@ -186,18 +173,6 @@ const Input: React.FC<
       />
     </div>
   );
-};
-
-/**
- * Converts a stat value to a tier
- */
-const statValueToTier = (value: number) => {
-  if (value >= 150) return "S";
-  if (value >= 120) return "A";
-  if (value >= 100) return "B";
-  if (value >= 80) return "C";
-  if (value >= 60) return "D";
-  return "F";
 };
 
 /**
